Add files via upload
This is the help file for the TestBLE application. This is a Python Tkinter based application to test BLE devices. It scans nearby BLE devices and then filters the Petasense Mote from the scanned devices. It automatically connects to the Mote device with the highest RSSI. Later it gives user option to change the SSID and Password as well on the connected device.